+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -10,10 +10,11 @@ import org.springframework.transaction.annotation.Transactional;
 import kitchenpos.order.dao.OrderRepository;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
-import kitchenpos.order.domain.OrderTable;
 import kitchenpos.order.dto.OrderLineItemRequest;
 import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.order.dto.OrderResponse;
+import kitchenpos.table.application.TableService;
+import kitchenpos.table.domain.OrderTable;
 
 @Service
 public class OrderService {
