+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -23,6 +23,8 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 import org.springframework.util.CollectionUtils;
 
+import kitchenpos.table.domain.OrderTable;
+
 @Entity
 @EntityListeners(AuditingEntityListener.class)
 public class Order {
