+++ src/main/java/kitchenpos/order/dto/OrderRequest.java
@@ -5,7 +5,7 @@ import java.util.List;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 
 public class OrderRequest {
     private Long orderTableId;
