+++ src/main/java/kitchenpos/table/dao/OrderTableRepository.java
@@ -1,10 +1,10 @@
-package kitchenpos.order.dao;
+package kitchenpos.table.dao;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 
 public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
