+++ src/main/java/kitchenpos/table/dao/TableGroupRepository.java
@@ -1,8 +1,8 @@
-package kitchenpos.order.dao;
+package kitchenpos.table.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.order.domain.TableGroup;
+import kitchenpos.table.domain.TableGroup;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
