+++ src/main/java/kitchenpos/table/dto/OrderTableRequest.java
@@ -1,6 +1,6 @@
-package kitchenpos.order.dto;
+package kitchenpos.table.dto;
 
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 
 public class OrderTableRequest {
     private Long tableGroupId;
