+++ src/main/java/kitchenpos/table/dto/TableGroupResponse.java
@@ -1,11 +1,11 @@
-package kitchenpos.order.dto;
+package kitchenpos.table.dto;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Objects;
 
-import kitchenpos.order.domain.OrderTable;
-import kitchenpos.order.domain.TableGroup;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.TableGroup;
 
 public class TableGroupResponse {
     private Long id;
