+++ src/test/java/kitchenpos/order/domain/OrdersTest.java
@@ -8,6 +8,8 @@ import java.util.Arrays;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import kitchenpos.table.domain.OrderTable;
+
 public class OrdersTest {
     
     @DisplayName("주문 목록에 주문을 추가한다")
